+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.0.0-alpha3] - In progress
+
+### Added
+
+- New `ConfigInterface::getArrayCopy()` method to have an array representation of configuration
+
 ## [2.0.0-alpha2] - 2021-03-12
 
 ### Changed